Refactoring the application, dividing responsibilities for the correct files
