+++ src/modules/accounts/repositories/IUsersRepository.ts
@@ -1,5 +1,6 @@
+import { User } from '@modules/accounts/infra/typeorm/entities/User';
+
 import { ICreateUserDTO } from '../dtos/ICreateUserDTO';
-import { User } from '../entities/User';
 
 interface IUsersRepository {
   create(data: ICreateUserDTO): Promise<void>;
