+++ src/modules/cars/infra/typeorm/repositories/CategoriesRepository.ts
@@ -1,10 +1,10 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { Category } from '../../entities/Category';
+import { Category } from '@modules/cars/infra/typeorm/entities/Category';
 import {
   ICategoriesRepository,
   ICreateCategoryDTO,
-} from '../ICategoriesRepository';
+} from '@modules/cars/repositories/ICategoriesRepository';
 
 class CategoriesRepository implements ICategoriesRepository {
   private repository: Repository<Category>;
