+++ src/modules/cars/infra/typeorm/repositories/SpecificationsRepository.ts
@@ -1,10 +1,10 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { Specification } from '../../entities/Specification';
+import { Specification } from '@modules/cars/infra/typeorm/entities/Specification';
 import {
   ISpecificationsRepository,
   ICreateSpecificationDTO,
-} from '../ISpecificationsRepository';
+} from '@modules/cars/repositories/ISpecificationsRepository';
 
 class SpecificationsRepository implements ISpecificationsRepository {
   private repository: Repository<Specification>;
