+++ src/modules/cars/repositories/ISpecificationsRepository.ts
@@ -1,4 +1,5 @@
-import { Specification } from '../entities/Specification';
+import { Specification } from '@modules/cars/infra/typeorm/entities/Specification';
+
 import { ICreateCategoryDTO } from './ICategoriesRepository';
 
 interface ICreateSpecificationDTO {
