+++ src/modules/cars/repositories/in-memory/CategoriesRepositoryInMemory.ts
@@ -1,4 +1,5 @@
-import { Category } from '../../entities/Category';
+import { Category } from '@modules/cars/infra/typeorm/entities/Category';
+
 import {
   ICategoriesRepository,
   ICreateCategoryDTO,
