+++ src/shared/infra/http/server.ts
@@ -4,13 +4,14 @@ import express, { Request, Response, NextFunction } from 'express';
 import 'express-async-errors';
 import swaggerUi from 'swagger-ui-express';
 
-import { AppError } from './errors/AppError';
+import { AppError } from '@shared/errors/AppError';
+
+import swaggerFile from '../../../swagger.json';
 import { router } from './routes';
-import swaggerFile from './swagger.json';
 
-import './database';
+import '@shared/infra/typeorm';
 
-import './shared/container';
+import '@shared/container';
 
 const app = express();
